refactor(#137/*): photo success code global/response 로 이동
